requirements.txt
Flask
python-dotenv
google-api-python-client
pandas
matplotlib
azure-ai-textanalytics
wordcloud
